minor improvements

+++ 09/storage/delete.php
@@ -1,15 +1,17 @@
 <?php
 
-// ha nincs azonosító paraméter irányítsuk vissza a felhasználót
-if (!isset($_GET['id'])) {
-    header('Location: index.php');
-    exit();
-}
-
 require_once 'ContactStorage.php';
 
-$contacts = new ContactStorage();
-
-$contacts->delete($_GET['id']);
+// ha nem post ne csináljunk semmit
+if($_SERVER['REQUEST_METHOD'] === 'POST') {
+	// ha nincs azonosító paraméter irányítsuk vissza a felhasználót
+	if (!isset($_GET['id'])) {
+		header('Location: index.php');
+		exit();
+	}
+	
+	$contacts = new ContactStorage();
+	$contacts->delete($_GET['id']);
+}
 
 header('Location: index.php');
